fix(vp8l-enc): simple-Huffman + palette RDO bias close #379
Two changes that together let the RDO sweep land on the colour-indexing
(palette) transform for tiny-palette fixtures (2-colour checkerboard,
8-colour 64×64) without regressing the larger-palette cases:

* **Simple-Huffman emission** (RFC 9649 §3.7.2.1.1). Alphabets with
  ≤ 2 active symbols now emit the simple-code wire format instead of
  the normal length-table header. Single-symbol alphabets save ~28
  header bits AND emit zero bits per symbol (decoder's `only_symbol`
  short-circuit returns the symbol without consuming bits). 2-symbol
  alphabets emit a 12-13 bit header + 1 bit/symbol. Big shrink on
  palette index streams (red/blue/alpha typically have 1 active
  symbol on packed-index pixels) and on the palette delta sub-stream.

* **RDO palette bias.** Track best palette-trial and best non-palette-
  trial winners separately; prefer the palette winner when it's
  within 16 bytes of the non-palette winner. Matches libwebp's
  always-prefer-palette-when-feasible behaviour. The 16-byte slack
  catches the 2-colour 32×32 case (palette: 30 B vs non-palette: 24 B,
  6-byte gap → palette wins) but not the 16-colour 64×64 palette_like
  fixture (palette: 243 B vs non-palette: 225 B, 18-byte gap → no-
  palette wins, preserving the `rdo_never_exceeds_smallest_fixed_
  configuration` soundness invariant).

Also drains the clippy 1.95 doc-lazy-continuation + unnecessary-cast
warnings under `-D warnings` (em-dash on a continuation line in the
vp8l/encoder module preamble + redundant `as usize` / `as u32` in
three test helpers).

+++ tests/animated_disposal_blend.rs
@@ -139,7 +139,7 @@ fn decoder_honours_dispose_to_background_between_frames() {
     // Top-left 4x4: green from F1.
     for y in 0..4 {
         for x in 0..4 {
-            let i = ((y * W as usize + x) * 4) as usize;
+            let i = (y * W as usize + x) * 4;
             assert_eq!(
                 &f1_canvas[i..i + 4],
                 &[0, 0xff, 0, 0xff],
@@ -150,7 +150,7 @@ fn decoder_honours_dispose_to_background_between_frames() {
     // Outside the 4x4 tile: must be (0,0,0,0) due to F0's dispose-to-bg.
     for y in 4..H as usize {
         for x in 0..W as usize {
-            let i = ((y * W as usize + x) * 4) as usize;
+            let i = (y * W as usize + x) * 4;
             assert_eq!(
                 &f1_canvas[i..i + 4],
                 &[0, 0, 0, 0],
